refactor: 패키지 구조 변경
http 패키지 내부로 주요 소스 이동
