+++ src/main/java/webserver/RequestHandler.java
@@ -7,6 +7,10 @@ import model.UserNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.IOUtils;
+import webserver.http.Request;
+import webserver.http.Response;
+import webserver.http.header.Header;
+import webserver.http.header.RequestHeader;
 
 import java.io.*;
 import java.net.Socket;
