+++ src/main/java/webserver/http/message/ResponseMessage.java
@@ -1,4 +1,8 @@
-package webserver;
+package webserver.http.message;
+
+import webserver.http.Body;
+import webserver.http.header.Header;
+import webserver.http.header.ResponseHeader;
 
 import java.util.Objects;
 import java.util.StringJoiner;
