--- src/test/java/webserver/HeaderTest.java
@@ -1,4 +1,0 @@
-package webserver;
-
-class HeaderTest {
-}
