+++ src/test/java/webserver/http/BodyTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import org.junit.jupiter.api.Test;
