+++ src/test/java/webserver/http/RequestTest.java
@@ -1,8 +1,12 @@
-package webserver;
+package webserver.http;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import webserver.http.header.RequestHeader;
+import webserver.http.message.GetMessage;
+import webserver.http.message.PostMessage;
+import webserver.http.message.RequestMessage;
 
 import java.util.Arrays;
 import java.util.HashMap;
