+++ src/test/java/webserver/http/ResponseTest.java
@@ -1,8 +1,9 @@
-package webserver;
+package webserver.http;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import webserver.http.message.ResponseMessage;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
