+++ src/test/java/webserver/http/header/HeaderTest.java
@@ -1,0 +1,4 @@
+package webserver.http.header;
+
+class HeaderTest {
+}
