+++ src/test/java/webserver/http/message/GetMessageTest.java
@@ -1,8 +1,10 @@
-package webserver;
+package webserver.http.message;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import webserver.http.header.Header;
+import webserver.http.header.RequestHeader;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,26 +12,25 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-class PostMessageTest {
-
+class GetMessageTest {
     @ParameterizedTest
     @MethodSource("getHeader")
     void getHeader(String messageText, RequestHeader expectedRequestHeader) {
-        PostMessage postMessage = PostMessage.from(messageText);
+        GetMessage getMessage = GetMessage.from(messageText);
 
-        assertThat(postMessage.getHeader()).isEqualTo(expectedRequestHeader);
+        assertThat(getMessage.getHeader()).isEqualTo(expectedRequestHeader);
     }
 
     static Stream<Arguments> getHeader() {
         return Stream.of(
+                //TODO getMessage인데 POST가  성공하는 테스트케이스로 들어가 있음!
                 Arguments.of("POST /user/create HTTP/1.1" + System.lineSeparator() +
                                 "Host: localhost:8080" + System.lineSeparator() +
                                 "Connection: keep-alive" + System.lineSeparator() +
                                 "Content-Length: 59" + System.lineSeparator() +
                                 "Content-Type: application/x-www-form-urlencoded" + System.lineSeparator() +
                                 "Accept: */*" + System.lineSeparator() +
-                                "" + System.lineSeparator() +
-                                "userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net" + System.lineSeparator(),
+                                "" + System.lineSeparator(),
                         Header.requestHeaderFrom("POST /user/create HTTP/1.1" + System.lineSeparator() +
                                 "Host: localhost:8080" + System.lineSeparator() +
                                 "Connection: keep-alive" + System.lineSeparator() +
@@ -42,50 +43,22 @@ class PostMessageTest {
     }
 
     @ParameterizedTest
-    @MethodSource("getBody")
-    void getBody(String messageText, Body expectedBody) {
-        PostMessage postMessage = PostMessage.from(messageText);
-
-        assertThat(postMessage.getBody()).isEqualTo(expectedBody);
-    }
-
-    static Stream<Arguments> getBody() {
-        return Stream.of(
-                Arguments.of("POST /user/create HTTP/1.1" + System.lineSeparator() +
-                                "Host: localhost:8080" + System.lineSeparator() +
-                                "Connection: keep-alive" + System.lineSeparator() +
-                                "Content-Length: 59" + System.lineSeparator() +
-                                "Content-Type: application/x-www-form-urlencoded" + System.lineSeparator() +
-                                "Accept: */*" + System.lineSeparator() +
-                                "" + System.lineSeparator() +
-                                "userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net" + System.lineSeparator(),
-                        Body.from("userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net" + System.lineSeparator())
-                ), Arguments.of("POST /user/create HTTP/1.1" + System.lineSeparator() +
-                                "" + System.lineSeparator() +
-                                "",
-                        Body.from("")
-                )
-        );
-    }
-
-    @ParameterizedTest
     @MethodSource("getMethod")
     void getMethod(String messageText, String expectedRequestMethod) {
-        PostMessage postMessage = PostMessage.from(messageText);
-        assertThat(postMessage.getMethod()).isEqualTo(expectedRequestMethod);
+        GetMessage getMessage = GetMessage.from(messageText);
+        assertThat(getMessage.getMethod()).isEqualTo(expectedRequestMethod);
     }
 
     static Stream<Arguments> getMethod() {
         return Stream.of(
-                Arguments.of("POST /user/create HTTP/1.1" + System.lineSeparator() +
+                Arguments.of("GET /user/create HTTP/1.1" + System.lineSeparator() +
                                 "Host: localhost:8080" + System.lineSeparator() +
                                 "Connection: keep-alive" + System.lineSeparator() +
                                 "Content-Length: 59" + System.lineSeparator() +
                                 "Content-Type: application/x-www-form-urlencoded" + System.lineSeparator() +
                                 "Accept: */*" + System.lineSeparator() +
-                                "" + System.lineSeparator() +
-                                "userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net",
-                        "POST"
+                                "" + System.lineSeparator(),
+                        "GET"
                 )
         );
     }
@@ -93,21 +66,21 @@ class PostMessageTest {
     @ParameterizedTest
     @MethodSource("getParameters")
     void getParameters(String messageText, Map<String, String> expectedParameters) {
-        PostMessage postMessage = PostMessage.from(messageText);
-        assertThat(postMessage.getParameters())
+        GetMessage getMessage = GetMessage.from(messageText);
+        assertThat(getMessage.getParameters())
                 .isEqualTo(expectedParameters);
     }
 
     static Stream<Arguments> getParameters() {
         return Stream.of(
-                Arguments.of("POST /user/create HTTP/1.1" + System.lineSeparator() +
+                Arguments.of(
+                        "GET /user/create?userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net HTTP/1.1" + System.lineSeparator() +
                                 "Host: localhost:8080" + System.lineSeparator() +
                                 "Connection: keep-alive" + System.lineSeparator() +
                                 "Content-Length: 59" + System.lineSeparator() +
                                 "Content-Type: application/x-www-form-urlencoded" + System.lineSeparator() +
                                 "Accept: */*" + System.lineSeparator() +
-                                "" + System.lineSeparator() +
-                                "userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net",
+                                "" + System.lineSeparator(),
                         new HashMap<String, String>() {{
                             put("userId", "javajigi");
                             put("password", "password");
