+++ src/test/java/webserver/http/message/RequestMessageTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http.message;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
