+++ src/test/java/webserver/http/message/ResponseMessageTest.java
@@ -1,8 +1,10 @@
-package webserver;
+package webserver.http.message;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import webserver.http.Body;
+import webserver.http.header.Header;
 
 import java.util.stream.Stream;
